refactor: Rename settings.py to config.py
